feat(mappings): Add Markdown mappings
Mappings have been added for bothe= the Markdown Preview plugin and the
Glow.nvim plugin, so that I no longer have to remember the commands. The
mapping for Markdown Preview might be changed at a later date to just
start the preview, as I tend to just close the browser instance and I'm
unsure if the current toggle map will keep the server running.